latest testing grounds
okay, so with this code we are able to identify the indices of every bit within every single bit of encoded ciphertext. If none of the ciphertext we wish to flip(i.e. the ";" and "=" characters) fall within the same index in a block, then by flipping the intial block of ciphertext at all necessary indices would all we would need to do